minor: Transfer `HelpCommand::toString` to `Utils`
