+++ src/Console/Command/HelpCommand.php
@@ -16,7 +16,7 @@ namespace PhpCsFixer\Console\Command;
 
 use PhpCsFixer\FixerConfiguration\AllowedValueSubset;
 use PhpCsFixer\FixerConfiguration\FixerOptionInterface;
-use PhpCsFixer\Preg;
+use PhpCsFixer\Utils;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\HelpCommand as BaseHelpCommand;
 use Symfony\Component\Console\Formatter\OutputFormatterStyle;
@@ -36,16 +36,6 @@ final class HelpCommand extends BaseHelpCommand
      * @var string
      */
     protected static $defaultName = 'help';
-
-    /**
-     * @param mixed $value
-     */
-    public static function toString($value): string
-    {
-        return \is_array($value)
-            ? self::arrayToString($value)
-            : self::scalarToString($value);
-    }
 
     /**
      * Returns the allowed values of the given option that can be converted to a string.
@@ -71,8 +61,8 @@ final class HelpCommand extends BaseHelpCommand
                 }
 
                 return strcasecmp(
-                    self::toString($valueA),
-                    self::toString($valueB)
+                    Utils::toString($valueA),
+                    Utils::toString($valueB)
                 );
             });
 
@@ -87,40 +77,5 @@ final class HelpCommand extends BaseHelpCommand
     protected function initialize(InputInterface $input, OutputInterface $output): void
     {
         $output->getFormatter()->setStyle('url', new OutputFormatterStyle('blue'));
-    }
-
-    /**
-     * @param mixed $value
-     */
-    private static function scalarToString($value): string
-    {
-        $str = var_export($value, true);
-
-        return Preg::replace('/\bNULL\b/', 'null', $str);
-    }
-
-    /**
-     * @param array<mixed> $value
-     */
-    private static function arrayToString(array $value): string
-    {
-        if (0 === \count($value)) {
-            return '[]';
-        }
-
-        $isHash = !array_is_list($value);
-        $str = '[';
-
-        foreach ($value as $k => $v) {
-            if ($isHash) {
-                $str .= self::scalarToString($k).' => ';
-            }
-
-            $str .= \is_array($v)
-                ? self::arrayToString($v).', '
-                : self::scalarToString($v).', ';
-        }
-
-        return substr($str, 0, -2).']';
     }
 }
