+++ src/Documentation/FixerDocumentGenerator.php
@@ -156,8 +156,8 @@ RST;
                     $allowedKind = 'Allowed values';
                     $allowed = array_map(static function ($value): string {
                         return $value instanceof AllowedValueSubset
-                            ? 'a subset of ``'.HelpCommand::toString($value->getAllowedValues()).'``'
-                            : '``'.HelpCommand::toString($value).'``';
+                            ? 'a subset of ``'.Utils::toString($value->getAllowedValues()).'``'
+                            : '``'.Utils::toString($value).'``';
                     }, $allowed);
                 }
 
@@ -165,7 +165,7 @@ RST;
                 $optionInfo .= "\n\n{$allowedKind}: {$allowed}";
 
                 if ($option->hasDefault()) {
-                    $default = HelpCommand::toString($option->getDefault());
+                    $default = Utils::toString($option->getDefault());
                     $optionInfo .= "\n\nDefault value: ``{$default}``";
                 } else {
                     $optionInfo .= "\n\nThis option is required.";
@@ -196,7 +196,7 @@ RST;
                     } else {
                         $doc .= sprintf(
                             "\n\nWith configuration: ``%s``.",
-                            HelpCommand::toString($sample->getConfiguration())
+                            Utils::toString($sample->getConfiguration())
                         );
                     }
                 }
@@ -238,7 +238,7 @@ RST;
 RST;
 
                 if (null !== $config) {
-                    $doc .= " with the config below:\n\n  ``".HelpCommand::toString($config).'``';
+                    $doc .= " with the config below:\n\n  ``".Utils::toString($config).'``';
                 } elseif ($fixer instanceof ConfigurableFixerInterface) {
                     $doc .= ' with the default config.';
                 } else {
