+++ src/Documentation/ListDocumentGenerator.php
@@ -116,8 +116,8 @@ RST;
                         $allowedKind = 'Allowed values';
                         $allowed = array_map(static function ($value): string {
                             return $value instanceof AllowedValueSubset
-                                ? 'a subset of ``'.HelpCommand::toString($value->getAllowedValues()).'``'
-                                : '``'.HelpCommand::toString($value).'``';
+                                ? 'a subset of ``'.Utils::toString($value->getAllowedValues()).'``'
+                                : '``'.Utils::toString($value).'``';
                         }, $allowed);
                     }
 
@@ -125,7 +125,7 @@ RST;
                     $documentation .= "\n     | {$allowedKind}: {$allowed}";
 
                     if ($option->hasDefault()) {
-                        $default = HelpCommand::toString($option->getDefault());
+                        $default = Utils::toString($option->getDefault());
                         $documentation .= "\n     | Default value: ``{$default}``";
                     } else {
                         $documentation .= "\n     | This option is required.";
