+++ tests/UtilsTest.php
@@ -379,6 +379,41 @@ final class UtilsTest extends TestCase
         self::assertNotContains($message, $triggered);
     }
 
+    /**
+     * @param mixed $input
+     *
+     * @dataProvider provideToStringCases
+     */
+    public function testToString(string $expected, $input): void
+    {
+        self::assertSame($expected, Utils::toString($input));
+    }
+
+    public static function provideToStringCases(): iterable
+    {
+        yield ["['a' => 3, 'b' => 'c']", ['a' => 3, 'b' => 'c']];
+
+        yield ['[[1], [2]]', [[1], [2]]];
+
+        yield ['[0 => [1], \'a\' => [2]]', [[1], 'a' => [2]]];
+
+        yield ['[1, 2, \'foo\', null]', [1, 2, 'foo', null]];
+
+        yield ['[1, 2]', [1, 2]];
+
+        yield ['[]', []];
+
+        yield ['1.5', 1.5];
+
+        yield ['false', false];
+
+        yield ['true', true];
+
+        yield ['1', 1];
+
+        yield ["'foo'", 'foo'];
+    }
+
     private function createFixerDouble(string $name, int $priority): FixerInterface
     {
         $fixer = $this->prophesize(FixerInterface::class);
